refactored core and output+log
