+++ staresc/core/worker/__init__.py
@@ -1,0 +1,2 @@
+from .scan import ScanWorker
+from .raw import RawWorker
